requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-mysql
apache-airflow-providers-apache-kafka
apache-airflow-providers-amazon
pandas
mysql-connector-python
faker
kafka-python
boto3
pulp